# KAFKA-8991: Enable scalac optimizer