--- a/gradle/spotbugs-exclude.xml
+++ b/gradle/spotbugs-exclude.xml
@@ -104,10 +104,8 @@ For a detailed description of spotbugs bug categories, see https://spotbugs.read
     </Match>
 
     <Match>
-        <!-- Suppression for the equals() for extensiom methods. -->
-        <Or>
-            <Class name="kafka.api.package$ElectLeadersRequestOps"/>
-        </Or>
+        <!-- Suppression for the equals() for extension methods. -->
+        <Class name="kafka.api.package$ElectLeadersRequestOps"/>
         <Bug pattern="EQ_UNUSUAL"/>
     </Match>
 
@@ -116,6 +114,24 @@ For a detailed description of spotbugs bug categories, see https://spotbugs.read
         <Source name="Json.scala"/>
         <Package name="kafka.utils"/>
         <Bug pattern="BC_VACUOUS_INSTANCEOF"/>
+    </Match>
+
+    <Match>
+        <!-- A spurious null check after inlining by the scalac optimizer confuses spotBugs -->
+        <Class name="kafka.log.Log"/>
+        <Bug pattern="NP_NULL_ON_SOME_PATH_EXCEPTION"/>
+    </Match>
+
+    <Match>
+        <!-- A spurious null check after inlining by the scalac optimizer confuses spotBugs -->
+        <Class name="kafka.tools.StateChangeLogMerger$"/>
+        <Bug pattern="NP_NULL_ON_SOME_PATH_FROM_RETURN_VALUE"/>
+    </Match>
+
+    <Match>
+        <!-- Unboxing to Int to make scalac happy makes spotBugs unhappy -->
+        <Class name="kafka.tools.ConsoleConsumer$ConsumerConfig"/>
+        <Bug pattern="BX_UNBOXING_IMMEDIATELY_REBOXED"/>
     </Match>
 
     <Match>