@lbradstreet 2019-10-07
I ran the JMH ReplicaFetcher benchmark from https://github.com/apache/kafka/pull/7443 with this change:

```
MBP 2015 2.5 GHz Intel Core i7

Before:
Benchmark                                  (partitionCount)  Mode  Cnt        Score       Error  Units
ReplicaFetcherThreadBenchmark.testFetcher               100  avgt   15    10463.937 ±   192.870  ns/op
ReplicaFetcherThreadBenchmark.testFetcher               500  avgt   15    72607.200 ±   522.900  ns/op
ReplicaFetcherThreadBenchmark.testFetcher              1000  avgt   15   165140.230 ±  1573.163  ns/op
ReplicaFetcherThreadBenchmark.testFetcher              5000  avgt   15  1382914.590 ± 17847.354  ns/op

After:
Benchmark                                  (partitionCount)  Mode  Cnt        Score       Error  Units
ReplicaFetcherThreadBenchmark.testFetcher               100  avgt   15    10207.193 ±   59.0646  ns/op
ReplicaFetcherThreadBenchmark.testFetcher               500  avgt   15    70814.440 ±   815.375  ns/op
ReplicaFetcherThreadBenchmark.testFetcher              1000  avgt   15   156053.360 ±  1153.965  ns/op
ReplicaFetcherThreadBenchmark.testFetcher              5000  avgt   15  1345613.534 ± 24347.230  ns/op
```