@ijuma 2019-10-09
2 jobs passed, Scala 2.11 failed with a single failing test but Scala 2.11 is not affected by this PR. @guozhangwang @omkreddy Does the PR look good to you?